i/prompting: move max id mmap to internal package
